requirements.txt
apache-airflow
scikit-learn
apache-airflow-providers-amazon
joblib
boto3
pandas
flask
